81: waitForTestsAreRunOnConsole
Task-Url: http://github.com/pitest/pitclipse/issues/81
