--- tests/org.pitest.pitclipse.ui.tests/src/org/pitest/pitclipse/ui/util/StepUtil.java
@@ -1,32 +1,0 @@
-/*******************************************************************************
- * Copyright 2012-2019 Phil Glover and contributors
- *  
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License.  You may obtain a copy
- * of the License at
- *  
- *  http://www.apache.org/licenses/LICENSE-2.0
- *  
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
- * License for the specific language governing permissions and limitations under
- * the License.
- ******************************************************************************/
-
-package org.pitest.pitclipse.ui.util;
-
-public class StepUtil {
-
-    private StepUtil() {
-    }
-
-    public static void safeSleep(long millis) {
-        try {
-            Thread.sleep(millis);
-        } catch (InterruptedException e) {
-            // Swallowed - no pretence of handling this. Unlikely to hit this in
-            // test code.
-        }
-    }
-}
